+++ app/src/main/java/de/baumann/quitsmoking/helper/DbAdapter_Files.java
@@ -21,14 +21,10 @@ package de.baumann.quitsmoking.helper;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.database.Cursor;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
-import android.preference.PreferenceManager;
-
-import de.baumann.quitsmoking.R;
 
 
 class DbAdapter_Files {